Split product type mutations into separate modules
