+++ saleor/graphql/product/mutations/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa: F401
+
+from .product_type_create import ProductTypeCreate
+from .product_type_delete import ProductTypeDelete
+from .product_type_update import ProductTypeUpdate
+
+__all__ = ["ProductTypeCreate" "ProductTypeUpdate" "ProductTypeDelete"]
